Changes to explanation.txt
Akira suggested a number of changes to explanation.txt to make the
document easier for a non-native English speaker to understand.  Along
with the wording changes, the letters used for events (A, B, C, D)
have been changed to (X, Y, Z, W) to avoid confusion with the word "A"
and CPU indentifiers.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Suggested-by: Akira Yokosawa <akiyks@gmail.com>